New options: Fs123DistribCache{Connect,Transfer}Timeout
The existing {Connect,Transfer}Timeout's remain in force when talking
to the origin server.  The new Peer{Connect,Transfer}Timeout's
apply when talking to a "peer" distributed cache.

If not explicitly specified, the Peer{Connect,Transfer}Timeouts will
be equal to whatever values are assigned (default or explicit) to the
old {Connect,Transfer}Timeout.

Change-Id: Id1eed087812cf69d6cc14bf49b9eabbdbbf9f111