Improve complaints on distrib-cache peer channel
Most problems encountered in the distrib-cache code are now
LOG_WARNING level.

Trouble on the server side results in two (not three or four!) groups
of LOG_WARNING complaints: one where the server complains directly,
and the other where the client sends the contents of the 5xx to the
complaint channel.

req::exception_reply in the fs123server library does *not* issue
a complaint.  It's up to the caller to do so (and to choose the
log-level, etc).

The complaint generated by receipt of an unexpected http status in
backend123_http::curl_handler::getreply wraps the contents of the
message between <curlhandler::getreply> </curlhandler::getreply> tags,
making it slightly easier to see what's going on.

Change-Id: I0245328fcd169ad7d4b9dc8809f7a987cb50ada8