Add an Fs123SquashAll and Fs123SquashRoot options
The default is empty, which means nothing is squashed

The format is either "~", or  (uid|uname)[:(gid|gname)]

Tilde means squash either all (SquashAll), or just root's (SquashRoot)
uids and gids to geteuid(), getegid(), respectively.  Otherwise,
squash uids and gids to the specified values (which may be either
numeric or string).  If the optional gid/gname is not specified, then
use the gid associated with the designated uid in the passwd database.

Change-Id: I4a6f7f60f8375a7c6be10de36a5445899525a5ef