Add a --bounded-max-age=<bool> (default true) server-side option
When a file has changed recently (i.e., it's st_mtime is in the recent
past), it's likely to change again in the near future.  So it's
desirable to place an upper bound of (now()-st_time) on the 'max-age'
specified in the reply's cache-control headers.

Similar logic was in the server before libfs123server was factored out.

Revive and enhance the t-14ccrules and ut_cc_rules unit tests, which
also went missing with the libfs123server refactor.  They both now
also test the new bounded-max-age behavior.

Other cosmetic and comment/doc-only changes, including deletion of
t-04changingcrf, which really is completely obsolete.

Change-Id: I1160923af67cffd0c160373bf22f129099873ad7