Fix intermittent failure of t-20touchexe
Out-of-order validators can happen with properly behaving clients and
servers if messages/threads "overtake" one another.  Before throwing
an error (ESTALE), the fs123_open callback retries the begetattr once
with no_cache=true, flushing the entry from the attrcache
and (hopefully) reaching all the way back to the origin for an
authoritative validator.

Abandon fs123_statistic_names and openfilemap_statistic_names in favor
of a single macro of statistic names in app_mount.hpp.

The 'max_stale' member of req123 is now an optional<int> (eliminating
the req123::MAX_STALE_UNSPECIFIED wart).  Eliminate the max_stale
argument to the req123::xxxreq "constructors".  max_stale and no_cache
are public data members with sensible defaults that the owner can
change as needed.

t-20touchexe makes a copy of the s/ and c/ directories in /tmp for
post-mortem analysis.

Change-Id: I7c60af9a6c9475b502ad551ec120dd7dfa428a84