Impose a maximum reply size in the server library
Check for too-large values when parsing the 'lenkib' part of request
URLs.

Thus,
  GET .../fs123/7/2/f/foo?10000000,0
no longer causes us to allocate a 10GB buffer!

There is now a:

   static const max_reply_size

in the fs123p7::req struct.  The value is hard-coded to 1025 kB, which
should be more than enough for any legitimate request from a fuse
client.

Change-Id: I31cea142fef3789ed9423a5f8ddd053ed56bbeb9