Major change - protocol 7.3 and new server API
This patch is the net result of several feature-branches
meged into the in-house DESRES version.

The 7.3 protocol removes http headers and reformats
the http message body as a list of key-value pairs.

Directory "offsets" are communicated as strings rather
than numbers, with corresponding changes to the fs123server
API.

New utilities and bug-fixes in core123.