Try to avoid locking up MacOS
On MacOS, if we successfully called fuse_mount, then *NO MATTER WHAT*,
we MUST go through with attaching a client to the mount point.  Most
of the new code is hidden behind #ifdef __APPLE__, but a tiny bit of
logic to detect whether fuse_session_loop ran to completion is in the
generic code path.