fix elastic_threadpool (again)
It must be the case that when submit() enqueues something, there is at
least one thread waiting to retire it.

A "proof" is elusive, but there are now soft_assert() at strategic
points in the code (none of which fire under heavy load).  It also
runs cleanly under the ThreadSanitizer and AddressSanitizer (despite
LWG issue 3343).  The unit test is much stricter, and fuzzes by
introducing random sleeps and yields into the control flow.  The
GNUmakefile runs ut_elastic_threadpool in parallel as part of 'make
check'

New features added to producerconsumerqueue.hpp

ut_expiring.cpp is tolerant of heavy load (e.g., when a bunch of
ut_elastic_threadpools are running along side it.)

The fs123server uses expiring_threadpool again.